+++ src/views/TagTimeline.vue
@@ -8,16 +8,18 @@
         :posts="posts"
         @load-next-page="loadNextPage"
       ></post-list>
+      <loader v-if="isLoading"></loader>
     </template>
   </sidebar-layout>
 </template>
 
 <script setup lang="ts">
-import { onMounted } from "vue"
+import { onMounted, ref } from "vue"
 import { $, $ref } from "vue/macros"
 import { useRoute } from "vue-router"
 
 import { Post, getTagTimeline } from "@/api/posts"
+import Loader from "@/components/Loader.vue"
 import PostList from "@/components/PostList.vue"
 import SidebarLayout from "@/components/SidebarLayout.vue"
 import { useCurrentUser } from "@/composables/user"
@@ -25,12 +27,15 @@ import { useCurrentUser } from "@/composables/user"
 const route = useRoute()
 const { authToken } = $(useCurrentUser())
 let posts = $ref<Post[]>([])
+const isLoading = ref(false)
 
 onMounted(async () => {
+  isLoading.value = true
   posts = await getTagTimeline(
     authToken,
     route.params.tagName as string,
   )
+  isLoading.value = false
 })
 
 async function loadNextPage(maxId: string) {
@@ -52,5 +57,9 @@ async function loadNextPage(maxId: string) {
   @include content-message;
 
   margin-bottom: $block-outer-padding;
+}
+
+.loader {
+  margin: $block-outer-padding auto;
 }
 </style>